nodejs: switch from EXTRA_OECONF to PACKAGECONFIG_CONFARGS in configure
Arguments from PACKAGECONFIG only get appended to EXTRA_OECONF if the
autotools class is inherited in the recipe. As this is not the case
with the node recipes, the PACKAGECONFIG settings never make it into the
build. Switch to directly using the PACKAGECONFIG_CONFARGS variable to
the right arguments get passed through.